+++ NEWS.md
@@ -5,6 +5,7 @@
 - Fixes
 
   - Fix possible cover generation memory leak. Undocumented cover RSS logger.
+  - Build a deadfall trap for search indexer zombies. Use Offspring's brains as bait.
 
 ## v0.10.1